requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
playwright==1.40.0
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.0
aiofiles==23.2.1 